refactor jwtencode and test
