+++ src/jwtencode.ts
@@ -1,20 +1,8 @@
-// @ts-check
-'use strict'
+import crypto from 'crypto'
 
-const crypto = require('crypto')
+import * as base64UrlSafe from './base64urlsafe'
 
-const base64UrlSafe = require('./base64urlsafe')
-
-function jwtEncode(privateKey, header, body, privateKeyPassword = null) {
-  if (
-    typeof header !== 'object' ||
-    Array.isArray(header) ||
-    typeof body !== 'object' ||
-    Array.isArray(body)
-  ) {
-    throw new Error('both header and body should be of type object')
-  }
-
+export function jwtEncode(privateKey: string, header: Record<string, unknown>, body: Record<string, unknown>, privateKeyPassword = null):  string {
   let signAlgo = null
   let hmacAlgo = null
   switch (header.alg) {
@@ -52,9 +40,9 @@ function jwtEncode(privateKey, header, body, privateKeyPassword = null) {
   }
 
   // Base64 encode header and body
-  let headerBase64 = base64UrlSafe.encode(Buffer.from(JSON.stringify(header)))
-  let bodyBase64 = base64UrlSafe.encode(Buffer.from(JSON.stringify(body)))
-  let headerBodyBase64 = headerBase64 + '.' + bodyBase64
+  const headerBase64 = base64UrlSafe.encode(Buffer.from(JSON.stringify(header)))
+  const bodyBase64 = base64UrlSafe.encode(Buffer.from(JSON.stringify(body)))
+  const headerBodyBase64 = headerBase64 + '.' + bodyBase64
 
   let signatureBuffer
   /* istanbul ignore else */
@@ -86,8 +74,6 @@ function jwtEncode(privateKey, header, body, privateKeyPassword = null) {
   }
 
   // Construct final JWT
-  let signatureBase64 = base64UrlSafe.encode(signatureBuffer)
+  const signatureBase64 = base64UrlSafe.encode(signatureBuffer)
   return headerBodyBase64 + '.' + signatureBase64
 }
-
-module.exports = jwtEncode
